+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [x] Update the text on onboarding phase 0 (Ad opt in request page)
 - [x] Agreeing to ads on the Onboarding page will now enable Ads on navigation by default. By default, this is every 4 page changes.
+- [x] Fix a bug in the server's SQL tables, where a image binary could potentially be too long to commit to the database.
 
 # 0.1.0+0617260921